refactor: move `handleSignOut` to `services`
